Add new request for adding new floors in to the system

+++ src/main/java/com/dstod/restaurantmanagerapi/management/models/dtos/CreateSectionResponse.java
@@ -1,12 +1,9 @@
 package com.dstod.restaurantmanagerapi.management.models.dtos;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
 public record CreateSectionResponse(
         Long id,
         String sectionName,
-        Integer floor,
+        Long floor,
         Boolean active
 ) {
 }
