+++ src/main/java/com/dstod/restaurantmanagerapi/management/models/dtos/SectionInfoDto.java
@@ -5,7 +5,7 @@ import java.util.List;
 public record SectionInfoDto(
         Long id,
         String sectionName,
-        Integer floor,
+        Long floor,
         Boolean active,
         List<TableInfoDto> tables
 ) {
